+++ example/main.tf
@@ -36,8 +36,9 @@ module "presto" {
   mode                  = "cluster"
   workers               = 1
   cluster_shared_secret = local.presto.cluster_shared_secret
-  consul_http_addr = "http://10.0.3.10:8500"
-  debug = true
+  consul_http_addr      = "http://10.0.3.10:8500"
+  debug                 = true
+  use_canary            = false
 
   minio         = local.minio
   hivemetastore = local.hivemetastore