+++ variables.tf
@@ -106,6 +106,12 @@ variable "container_environment_variables" {
   default     = [""]
 }
 
+variable "use_canary" {
+  type = bool
+  description = "Uses canary deployment for Presto"
+  default = true
+}
+
 variable "hivemetastore" {
   type = object({
     service_name = string,