+++ example/main.tf
@@ -84,13 +84,13 @@ module "postgres" {
   nomad_namespace   = local.nomad_namespace
 
   # postgres
-  postgres_service_name                    = "postgres"
-  postgres_container_image                 = "postgres:12-alpine"
-  postgres_container_port                  = 5432
-  postgres_admin_user                      = "hive"
-  postgres_admin_password                  = "hive"
-  postgres_database                        = "metastore"
-  postgres_container_environment_variables = ["PGDATA=/var/lib/postgresql/data"]
+  service_name                    = "postgres"
+  container_image                 = "postgres:12-alpine"
+  container_port                  = 5432
+  admin_user                      = "hive"
+  admin_password                  = "hive"
+  database                        = "metastore"
+  container_environment_variables = ["PGDATA=/var/lib/postgresql/data"]
 }
 
 module "hive" {