+++ CHANGELOG.md
@@ -8,12 +8,14 @@
 - Documentation #3 #8
 - Fixate linter version #10
 - Consul Connect enabled multi node cluster #13 #14 #16 #19 #24
+- Code to support successful execution of nomad presto job and tests when consul_acl_default_policy is deny #32
 
 ## Changed
 
 - Use docker instead fo binaries #21
 - size vagrant box to runner in github actions #23
 - Sync origin template #28
+- Bumped version for terraform-minio-module 0.0.3 -> 0.1.0 #34
 
 ## Fixed