+++ CHANGELOG.md
@@ -17,7 +17,9 @@
 - Use docker instead fo binaries #21
 - size vagrant box to runner in github actions #23
 - Sync origin template #28
-- Bumped version for minio, hive and postgres #34
+- Bumped version minio: 0.0.2 -> 0.1.0
+- Bumped version hive: 0.0.2 -> 0.1.0
+- Bumped version postgres: 0.0.1 -> 0.1.0
 - Consul Connect plugin optional with configurable artifact source #30
 
 ## Fixed